refactor(core): address review feedback — fix layering and preserve output
- Move build_escalation_question to core/fix_tracker.py (neutral location),
  fixing the layering concern where core/react_agent imported from adapters
- Preserve adapter last output in blocker results (was silently dropped)
- Consolidate test imports
